add /admin

+++ src/main/kotlin/streetlight/server/db/ApiRouting.kt
@@ -2,7 +2,6 @@ package streetlight.server.db
 
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.call
-import io.ktor.server.auth.authenticate
 import io.ktor.server.request.receive
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Routing
@@ -11,8 +10,7 @@ import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 import io.ktor.server.routing.put
 import org.jetbrains.exposed.dao.IntEntity
-import streetlight.server.plugins.authenticateJwt
-import streetlight.server.plugins.v1
+import streetlight.server.plugins.*
 
 inline fun <reified Data : Any, DataEntity : IntEntity> Routing.applyServiceRouting(
     endpoint: String, service: DataService<Data, DataEntity>
@@ -40,12 +38,20 @@ inline fun <reified Data : Any, DataEntity : IntEntity> Routing.applyServiceRout
 
     authenticateJwt {
         post("$v1/${endpoint}") {
+            if (!call.testRole(ROLE_ADMIN)) {
+                call.respond(HttpStatusCode.Forbidden)
+                return@post
+            }
             val data = call.receive<Data>()
             val id = service.create(data)
             call.respond(HttpStatusCode.Created, id)
         }
 
         put("$v1/${endpoint}/{id}") {
+            if (!call.testRole(ROLE_ADMIN)) {
+                call.respond(HttpStatusCode.Forbidden)
+                return@put
+            }
             val id = call.getIdOrThrow()
             val data = call.receive<Data>()
             service.update(id, data)
@@ -53,6 +59,10 @@ inline fun <reified Data : Any, DataEntity : IntEntity> Routing.applyServiceRout
         }
 
         delete("$v1/${endpoint}/{id}") {
+            if (!call.testRole(ROLE_ADMIN)) {
+                call.respond(HttpStatusCode.Forbidden)
+                return@delete
+            }
             val id = call.getIdOrThrow()
             service.delete(id)
             call.respond(HttpStatusCode.OK)
